requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.5.0
websockets==12.0
python-dotenv==1.0.0
slowapi==0.1.9
psycopg[binary]==3.2.3
